Move projects metadata for SBOM export to Google Spreadsheet
Move the "known" project's metadata to a metadata spreadsheet where
we will be able to maintain it without having to modify the code in
breeze. This allows to regenerate SBOM export any time with updated
metadata and allows it to be edited and maintained manually.

This also means that the CSV option should be removed (since we are
now rely on Google Spreadsheet anyway). The code has been refactored
to move things around betwen cdxgen/sbom/ne projects_google_spreadsheet
modules.

Added rich progress to track the advancements of tasks.
