+++ dev/breeze/doc/images/output_sbom_export-dependency-information.txt
@@ -1,1 +1,1 @@
-b66a78c14ba8a90c256d630d5246fc5d
+9b70343c1f234949268e2a38a1b9fef4
